Run clang-format and fix up cpplint issues.

+++ src/ofp/constants.cpp
@@ -9,7 +9,7 @@ UInt16 ofp::OFPGetDefaultPort() {
   if (!portStr)
     return OFP_DEFAULT_PORT;
 
-  long portNum = std::strtol(portStr, NULL, 10);
+  int portNum = static_cast<int>(std::strtol(portStr, NULL, 10));
   if (portNum > 0 && portNum < 0xffff) {
     return UInt16_narrow_cast(portNum);
   }
